+++ StabilityMatrix.Avalonia/App.axaml.cs
@@ -259,7 +259,8 @@ public sealed class App : Application
             .AddSingleton<LaunchPageViewModel>()
             .AddSingleton<ProgressManagerViewModel>()
             .AddSingleton<InferenceViewModel>()
-            .AddSingleton<ProgressManagerViewModel>();
+            .AddSingleton<ProgressManagerViewModel>()
+            .AddSingleton<OutputsPageViewModel>();
 
         services.AddSingleton<MainWindowViewModel>(
             provider =>
@@ -278,6 +279,7 @@ public sealed class App : Application
                         provider.GetRequiredService<PackageManagerViewModel>(),
                         provider.GetRequiredService<CheckpointsPageViewModel>(),
                         provider.GetRequiredService<CheckpointBrowserViewModel>(),
+                        provider.GetRequiredService<OutputsPageViewModel>()
                     },
                     FooterPages = { provider.GetRequiredService<SettingsViewModel>() }
                 }
@@ -388,6 +390,7 @@ public sealed class App : Application
         services.AddSingleton<ProgressManagerPage>();
         services.AddSingleton<InferencePage>();
         services.AddSingleton<NewCheckpointsPage>();
+        services.AddSingleton<OutputsPage>();
 
         // Inference tabs
         services.AddTransient<InferenceTextToImageView>();
@@ -430,12 +433,13 @@ public sealed class App : Application
     internal static void ConfigurePackages(IServiceCollection services)
     {
         services.AddSingleton<BasePackage, A3WebUI>();
-        services.AddSingleton<BasePackage, Fooocus>();
+        services.AddSingleton<BasePackage, StableDiffusionUx>();
         services.AddSingleton<BasePackage, VladAutomatic>();
-        services.AddSingleton<BasePackage, InvokeAI>();
+        services.AddSingleton<BasePackage, Fooocus>();
+        services.AddSingleton<BasePackage, FooocusMre>();
         services.AddSingleton<BasePackage, ComfyUI>();
         services.AddSingleton<BasePackage, VoltaML>();
-        services.AddSingleton<BasePackage, FooocusMre>();
+        services.AddSingleton<BasePackage, InvokeAI>();
     }
 
     private static IServiceCollection ConfigureServices()