+++ StabilityMatrix.Avalonia/Views/OutputsPage.axaml.cs
@@ -1,0 +1,14 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+using StabilityMatrix.Avalonia.Controls;
+
+namespace StabilityMatrix.Avalonia.Views;
+
+public partial class OutputsPage : UserControlBase
+{
+    public OutputsPage()
+    {
+        InitializeComponent();
+    }
+}