+++ StabilityMatrix.Core/Models/Packages/UnknownPackage.cs
@@ -26,6 +26,8 @@ public class UnknownPackage : BasePackage
 
     public override SharedFolderMethod RecommendedSharedFolderMethod => SharedFolderMethod.Symlink;
 
+    public override string OutputFolderName { get; }
+
     public override Task DownloadPackage(
         string installLocation,
         DownloadPackageVersionOptions versionOptions,
@@ -83,6 +85,16 @@ public class UnknownPackage : BasePackage
         throw new NotImplementedException();
     }
 
+    public override Task SetupOutputFolderLinks(DirectoryPath installDirectory)
+    {
+        throw new NotImplementedException();
+    }
+
+    public override Task RemoveOutputFolderLinks(DirectoryPath installDirectory)
+    {
+        throw new NotImplementedException();
+    }
+
     public override IEnumerable<TorchVersion> AvailableTorchVersions =>
         new[] { TorchVersion.Cuda, TorchVersion.Cpu, TorchVersion.Rocm, TorchVersion.DirectMl };
 
@@ -121,6 +133,12 @@ public class UnknownPackage : BasePackage
         Task.FromResult(Enumerable.Empty<Release>());
 
     public override List<LaunchOptionDefinition> LaunchOptions => new();
+
+    public override Dictionary<SharedFolderType, IReadOnlyList<string>>? SharedFolders { get; }
+    public override Dictionary<
+        SharedOutputType,
+        IReadOnlyList<string>
+    >? SharedOutputFolders { get; }
 
     public override Task<string> GetLatestVersion() => Task.FromResult(string.Empty);